requirements.txt
loguru==0.6.0
pydantic==1.10.15
PyJWT==2.8.0
PyMySQL==1.1.0
PyYAML==6.0.1
SQLAlchemy==1.4.52

fastapi~=0.92.0
openai~=1.16.2
starlette~=0.25.0
requests~=2.31.0
uvicorn~=0.29.0
httpx~=0.27.0
python-dotenv~=1.0.1
pyhanlp~=0.1.85
Jinja2~=3.1.3
sse-starlette~=2.0.0
redis~=5.0.4
pocode~=0.0.3